+++ docs/release-notes.md
@@ -11,6 +11,7 @@
 * For modders:
   * Added code analysis to mod build config package to flag common issues as warnings.
   * Added `Context.IsMultiplayer` and `Context.IsMainPlayer` flags.
+  * Added `Constants.TargetPlatform` which says whether the game is running on Linux, Mac, or Windows.
   * Fixed assets loaded by temporary content managers not being editable by mods.
   * Fixed assets not reloaded consistently when the player switches language.
   * Fixed console command input not saved to the log.